+++ resources/assets/js/pages/administration/users/Create.vue
@@ -5,7 +5,7 @@
             <vue-form-ss class="box animated fadeIn"
                 :params="[$route.name, null, false]">
                 <template slot="owner_id" slot-scope="{ field, errors }">
-                    <vue-select name="owner_id"
+                    <vue-select :i18n="__"
                         v-model="field.value"
                         :has-error="errors.has(field.name)"
                         @input="pivotParams.owners.id=$event;errors.clear(field.name)"
@@ -13,7 +13,7 @@
                     </vue-select>
                 </template>
                 <template slot="role_id" slot-scope="{ field, errors }">
-                    <vue-select name="role_id"
+                    <vue-select :i18n="__"
                         :pivot-params="pivotParams"
                         v-model="field.value"
                         :has-error="errors.has(field.name)"
@@ -29,6 +29,7 @@
 
 <script>
 
+import { mapGetters } from 'vuex';
 import VueFormSs from '../../../components/enso/vueforms/VueFormSs.vue';
 import VueSelect from '../../../components/enso/select/VueSelect.vue';
 
@@ -39,6 +40,10 @@ export default {
         return {
             pivotParams: { owners: { id: null } },
         };
+    },
+
+    computed: {
+        ...mapGetters('locale', ['__']),
     },
 };